refactor: Commentsコンポーネントに分割した
